requirements.txt
openai>=1.0.0
flask>=3.0.0
requests>=2.31.0
python-whois>=0.9.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=4.9.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
